backend/requirements.txt
asgiref==3.7.2
attrs==23.2.0
Django==4.2.5
djangorestframework==3.14.0
drf-spectacular==0.27.1
factory-boy==3.3.0
Faker==22.6.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
packaging==23.2
pillow==10.2.0
pluggy==1.4.0
psycopg2-binary==2.9.7
pytest==8.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.17.1
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
